tvg helper: adding predefined variables, changing flag/indicator names
The size of the flag, indicator and byte counter has been predefined.
Names of flags and indicators has been changed. The indicator is used
to inform what data is coming after it, the flag is used as the data type
for enum values.
Part of the code was moved from thorvg.h to tvgTvgHelper.h.